requirements.txt
aiogram==2.14.3
psycopg2-binary==2.9.1